Add files via upload
This Java-based library management system is a console application created as a learning project while exploring Java OOP concepts. The application provides essential functionalities such as adding, searching, updating, deleting, and displaying books. With a simple menu-driven interface, it serves as a hands-on exercise in applying object-oriented principles in Java.